update: 도메인 기반 구조 정리
