+++ src/main/java/io/github/bon/wonx/domain/video/entity/Video.java
@@ -1,5 +1,6 @@
-package io.github.bon.wonx.domain.video;
+package io.github.bon.wonx.domain.video.entity;
 
+import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
 import lombok.Getter;
@@ -22,5 +23,5 @@ public class Video {
   private String ageRating; // 시청 등급
   private String backgroundVideoUrl; // 배너 등에 쓰일 영상 배경 URL
   private String thumbnailUrl; // 썸네일 이미지 URL
-
+  private Integer viewCount; // 콘텐츠 조회수
 }
