+++ src/main/java/io/github/bon/wonx/domain/video/service/VideoService.java
@@ -1,23 +1,31 @@
-package io.github.bon.wonx.service;
+package io.github.bon.wonx.domain.video.service;
 
-import io.github.bon.wonx.domain.video.Video;
-import io.github.bon.wonx.domain.video.VideoRepository;
+import io.github.bon.wonx.domain.video.entity.Video;
+import io.github.bon.wonx.domain.video.repository.VideoRepository;
+import org.springframework.data.domain.Pageable;
+
+import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDate;
 import java.util.List;
 
 @Service
-public class MainService {
+public class VideoService {
   private final VideoRepository videoRepository;
 
-  public MainService(VideoRepository videoRepository) {
+  public VideoService(VideoRepository videoRepository) {
     this.videoRepository = videoRepository;
   }
 
   public Video getMainBannerVideo() {
     List<Video> topVideo = videoRepository.findTop1ByOrderByRatingDesc();
     return topVideo.isEmpty() ? null : topVideo.get(0);
+  }
+
+  public List<Video> getPopularVideos(int count) {
+    Pageable pageable = PageRequest.of(0, count);
+    return videoRepository.findAllByOrderByViewCountDesc(pageable);
   }
 
   public List<Video> getUpcomingVideosWithinAWeek() {
